test: move the camera mercator coordinate tests to mercator_utils
