+++ src/geo/projection/mercator_utils.test.ts
@@ -1,8 +1,10 @@
 import {describe, expect, test} from 'vitest';
 import Point from '@mapbox/point-geometry';
 import {LngLat} from '../lng_lat.ts';
-import {getMercatorHorizon, projectToWorldCoordinates, tileCoordinatesToLocation, tileCoordinatesToMercatorCoordinates} from './mercator_utils.ts';
+import {cameraMercatorCoordinate, getMercatorHorizon, projectToWorldCoordinates, tileCoordinatesToLocation, tileCoordinatesToMercatorCoordinates} from './mercator_utils.ts';
 import {MercatorTransform} from './mercator_transform.ts';
+import {GlobeTransform} from './globe_transform.ts';
+import {altitudeFromMercatorZ} from '../mercator_coordinate.ts';
 import {CanonicalTileID} from '../../tile/tile_id.ts';
 import {EXTENT} from '../../data/extent.ts';
 import {createIdentityMat4f32, MAX_VALID_LATITUDE} from '../../util/util.ts';
@@ -47,6 +49,53 @@ describe('mercator utils', () => {
 
         expect(horizon).toBeCloseTo(-75.52102888757743, 10);
     });
+    describe('cameraMercatorCoordinate', () => {
+        test('places the camera above the ground', () => {
+            const transform = new MercatorTransform({minZoom: 0, maxZoom: 22, minPitch: 0, maxPitch: 180, renderWorldCopies: true});
+            transform.setElevation(200);
+            transform.setCenter(new LngLat(15.0, 55.0));
+            transform.setZoom(14);
+            transform.setPitch(55);
+            transform.setBearing(75);
+            transform.resize(512, 512);
+
+            const cameraCoord = cameraMercatorCoordinate(transform);
+
+            expect(cameraCoord.toLngLat().lng).toBeCloseTo(14.973921529405033, 10);
+            expect(cameraCoord.toLngLat().lat).toBeCloseTo(54.99599181678275, 10);
+            expect(altitudeFromMercatorZ(cameraCoord.z, cameraCoord.y)).toBeCloseTo(transform.getCameraAltitude(), 0);
+        });
+
+        test('does not depend on the projection', () => {
+            const precisionDigits = 10;
+            const globeTransform = new GlobeTransform();
+            globeTransform.resize(512, 512);
+            globeTransform.setZoom(5);
+            globeTransform.setCenter(new LngLat(15, 55));
+            globeTransform.setMaxPitch(60);
+            globeTransform.setPitch(55);
+            globeTransform.setBearing(75);
+
+            const mercatorTransform = new MercatorTransform({minZoom: 0, maxZoom: 22, minPitch: 0, maxPitch: 60, renderWorldCopies: true});
+            mercatorTransform.resize(512, 512);
+            mercatorTransform.setZoom(5);
+            mercatorTransform.setCenter(new LngLat(15, 55));
+            mercatorTransform.setPitch(55);
+            mercatorTransform.setBearing(75);
+
+            const cameraCoord = cameraMercatorCoordinate(globeTransform);
+            const mercatorCameraCoord = cameraMercatorCoordinate(mercatorTransform);
+
+            expect(cameraCoord.x).toBeCloseTo(mercatorCameraCoord.x, precisionDigits);
+            expect(cameraCoord.y).toBeCloseTo(mercatorCameraCoord.y, precisionDigits);
+            expect(cameraCoord.z).toBeCloseTo(mercatorCameraCoord.z, precisionDigits);
+
+            // The pixel under a pitched globe camera misses the sphere, so the screen round trip lands elsewhere.
+            const screenRoundTrip = globeTransform.screenPointToMercatorCoordinate(globeTransform.getCameraPoint());
+            expect(cameraCoord.toLngLat().distanceTo(screenRoundTrip.toLngLat())).toBeGreaterThan(50000);
+        });
+    });
+
     describe('getProjectionData', () => {
         test('return identity matrix when not passing overscaledTileID', () => {
             const transform = new MercatorTransform({minZoom: 0, maxZoom: 22, minPitch: 0, maxPitch: 180, renderWorldCopies: true});
